Modified import.clj to map all premises to ordinary premises on import.
This is the first step towards simplifying Carneades.  In arguments,
all premises will be ordinary.  The distinction between ordinary
premises, exceptions and assumptions is preserved in argumentation
schemes.  In addition, the "scintilla of evidence" (SE) proof standard
will be removed.  Statements assigned the SE standard in LKIF files
will be reassigned the "preponderance of evidence" (PE) standard on
import.