Reified premises; Defined an abstract "literal" type, covering both statements and s-expressions; Allow sections of theories to be nested.
- Premises are now reified as records, to allow additional properties
to be defined, such as whether or not the premise was "implicit" in the
text of a reconstructed argument. (The "implicit" property was
requested by Doug Walton.)

- Statements and expression representations can now both be used in
most places requiring literals.  An abstract literal type has been
defined, covering both statements and s-expressions.  This simplifies
the code in many places, since it is usually no longer necessary to
translate between s-expressions and statements.

- Sections of theories can now contains (sub)sections, allowing a
theory to be organized in a hierarchical structure.  Schemes can be
defined directly at each level of the hierarchy, including the top
level of a theory, so that sections are now optional.